requirements.txt
kafka-python==2.0.2
python-telegram-bot==20.7
python-dotenv==1.0.0
confluent-kafka==2.3.0
prometheus-client==0.19.0
requests==2.31.0
